+++ components/UI/Nav.tsx
@@ -7,7 +7,7 @@ import useLoginStore from "@/store/useLoginStore";
 import { FaBell } from "react-icons/fa";
 import DefaultProfile from "./DefaultProfile";
 import AdminNav from "./AdminNav";
-import Notice from "../Notification/Notice";
+import App from "@/app/page";
 import { RxHamburgerMenu } from "react-icons/rx";
 import { useState } from "react";
 import MobileMenu from "./MoblieMenu";
@@ -88,7 +88,7 @@ export default function Nav() {
           <button>
             <FaBell className="h-6 w-10 text-blue-400" onClick={toggleModal} />
           </button>
-          {modal && <Notice />}
+          {modal && <App />}
 
           {loginState ? (
             <div className="dropdown dropdown-end">
